updated example treatment plan files
Example plan now has organ contour volumes closer to what is encountered clinically. Fulfilling the following comment in the readme.md file:
"The example treatment plan was created from a phantom with a small prostate volume. The small volume causes large variations in DVH metrics as uncertainty shifts are applied to needles and other structures. Although this plan provides an adequate test plan to explore the features of RobustBrachy, the intention soon is to provide an example plan with a prostate volume that more exemplifies the average patient undergoing HDR prostate brachytherapy."